source/backend/requirements.txt
Django>=3.2
djangorestframework
psycopg2-binary
djangorestframework_simplejwt
django-cors-headers
django-cryptography
pandas>=1.3.0
django-filter
drf-nested-routers==0.93.4
openai
chardet==5.2.0
requests
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
Pillow==10.1.0
python-docx==1.1.0
PyMuPDF==1.23.8
PyPDF2==3.0.1
# Task Scheduling
croniter>=1.4.1
django-extensions
pytest>=7.0.0
pytest-django>=4.5.0
stripe>=8.0.0
pytest-cov
factory-boy
pytest-xdist
pytest-mock